Add xdl-jax package, TBSM model, and XQT GEMM/SVD-FLUX runtime expansions
- xdl-jax: new JAX training subproject (config, registry, callbacks, checkpoint, distributed) with CI workflows and docs
- XDL: TBSM generate model and trainer integration
- XQT GEMM: w4a8 sm89, AWQ w4a16 decode sm89, sm90 FP8 WGMMA CUDA kernel, sm120 NVFP4 backend, lowbit contracts and tests
- XQT runtime: SVD FLUX attention/block/transformer and GELU MLP composite modules, convrot int8 quantizer updates
- XQT analysis: visualization module
- research: SVDQ official parity benchmarks plus curated norm-runner and MLP result records
- docs: XQT workflow, xdl-jax architecture, operator optimization records

+++ xqt/operator_opt/kernels/triton/pointwise.py
@@ -128,10 +128,10 @@ if triton is not None and tl is not None:
         offsets = tl.arange(0, BLOCK_SIZE)
         mask = offsets < hidden_dim
         base = row_id * hidden_dim + offsets
-        x = tl.load(x_ptr + base, mask=mask, other=0.0)
+        x = tl.load(x_ptr + base, mask=mask, other=0.0).to(tl.float32)
         square_sum = tl.sum(x * x, axis=0)
         rms = tl.rsqrt(square_sum / hidden_dim + eps)
-        weight = tl.load(weight_ptr + offsets, mask=mask, other=0.0)
+        weight = tl.load(weight_ptr + offsets, mask=mask, other=0.0).to(tl.float32)
         out = x * rms * weight
         tl.store(out_ptr + base, out, mask=mask)
 
@@ -150,14 +150,14 @@ if triton is not None and tl is not None:
         offsets = tl.arange(0, BLOCK_SIZE)
         mask = offsets < hidden_dim
         base = row_id * hidden_dim + offsets
-        merged = tl.load(x_ptr + base, mask=mask, other=0.0) + tl.load(
+        merged = tl.load(x_ptr + base, mask=mask, other=0.0).to(tl.float32) + tl.load(
             residual_ptr + base,
             mask=mask,
             other=0.0,
-        )
+        ).to(tl.float32)
         square_sum = tl.sum(merged * merged, axis=0)
         rms = tl.rsqrt(square_sum / hidden_dim + eps)
-        weight = tl.load(weight_ptr + offsets, mask=mask, other=0.0)
+        weight = tl.load(weight_ptr + offsets, mask=mask, other=0.0).to(tl.float32)
         out = merged * rms * weight
         tl.store(out_ptr + base, out, mask=mask)
